+++ src/otx/core/model/entity/action_classification.py
@@ -41,27 +41,9 @@ class MMActionCompatibleModel(OTXActionClsModel):
         super().__init__(num_classes=num_classes)
 
     def _create_model(self) -> nn.Module:
-        from mmaction.models.data_preprocessors import (
-            ActionDataPreprocessor as _ActionDataPreprocessor,
-        )
-        from mmaction.registry import MODELS
-        from mmengine.registry import MODELS as MMENGINE_MODELS
-
-        # NOTE: For the history of this monkey patching, please see
-        # https://github.com/openvinotoolkit/training_extensions/issues/2743
-        @MMENGINE_MODELS.register_module(force=True)
-        class ActionDataPreprocessor(_ActionDataPreprocessor):
-            @property
-            def device(self) -> device:
-                try:
-                    buf = next(self.buffers())
-                except StopIteration:
-                    return super().device
-                else:
-                    return buf.device
-
-        self.classification_layers = get_classification_layers(self.config, MODELS, "model.")
-        return build_mm_model(self.config, MODELS, self.load_from)
+        from .utils.mmaction import create_model
+        model, self.classification_layers = create_model(self.config, self.load_from)
+        return model
 
     def _customize_inputs(self, entity: ActionClsBatchDataEntity) -> dict[str, Any]:
         """Convert ActionClsBatchDataEntity into mmaction model's input."""
