+++ src/otx/core/model/entity/utils/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (C) 2024 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+#
+"""Utility files for model entities used in OTX."""
